raspberry_ui/requirements.txt
flask==2.2.2
pyserial